requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
boto3>=1.26.0
numpy>=1.21.0
Pillow>=9.0.0
pyyaml>=6.0
tqdm>=4.64.0
scipy>=1.7.0
